Bugfix in EvenlySignal#resample and EvenlySignal#get_x_values, refactoring
Memory improvement in the fast downsampling.

An error was thrown in the case of the non fast interpolation because of the length of the indexes_out array. I doubled the last sample of the signal and edited the indexes_out generation as the ratio round error sum up using arange with step and the final length could be wrong.

Renamed UnevenlyTimeSignal to sparse signal.